backend/requirements.txt
django==5.1.1
django-environ==0.11.2
pymysql==1.1.1
cryptography==42.0.5
